+++ CHANGELOG.md
@@ -3,6 +3,13 @@
 This document provides information about the changes and new features in Tlumach.
 
 ---
+Version: 1.0.1  
+Date: November 26, 2025
+
+- [FIX] Fixed loading of default translation files from a subdirectory, when both the config file and the translation file resided in the same _sub_directory.
+- [FIX] TOML parser falsefully marked some units as templated.
+
+---
 Version: 1.0.0  
 Date: November 26, 2025